+++ AtlasLootClassic/Documentation/Release_Notes.md
@@ -1,3 +1,7 @@
+# v4.2.1 (Nov 4, 2024)
+- Added missing trinkets to Fissure Stone vendor
+- Fixed dungeon sets not showing under classic collections
+
 # v4.2.0 (Oct 29, 2024)
 - Fixed model display, shows models again (some models buggy)
 - Added Phase 3 VP/JP items
@@ -8,4 +12,3 @@
 - Many small fixes and code cleanup
 
 Shoutout to @Firsthunterz for help adding lots of items manually in this release
-